+++ src/Test/WebDriver/Commands/DocumentHandling.hs
@@ -11,10 +11,10 @@ import Data.Aeson as A
 import Data.CallStack
 import qualified Data.Foldable as F
 import Data.Text (Text)
+import Test.WebDriver.Exceptions
+import Test.WebDriver.JSON
 import Test.WebDriver.Types
 import Test.WebDriver.Util.Commands
-import Test.WebDriver.Exceptions.Internal
-import Test.WebDriver.JSON
 import UnliftIO.Exception (handle, throwIO)
