+++ src/Test/WebDriver/Commands/Navigation.hs
@@ -13,10 +13,10 @@ import Data.CallStack
 import Data.Text (Text)
 import Network.URI hiding (path)  -- suppresses warnings
 import Prelude -- hides some "unused import" warnings
+import Test.WebDriver.Exceptions
+import Test.WebDriver.JSON
 import Test.WebDriver.Types
 import Test.WebDriver.Util.Commands
-import Test.WebDriver.Exceptions.Internal
-import Test.WebDriver.JSON
 import UnliftIO.Exception (throwIO)
